Add nickname screen and update navigation; refactor AuthPage for email/password sign-in
